+++ CHANGELOG.md
@@ -19,7 +19,8 @@
 - Added confirm dialog on item's deletion (Hold "ctrl" on the click, if you want to bypass it)
 - Added "Sleep" & "Scene End" buttons in "GM ToolBox" (old "difficulty" box)
 - Added an option to reverse the token's bar on fatigue (thanks to Jzrzmy), and colorize in red the strife bar if compromise
-- Split techniques in actor sheet for better readability
+- Split Techniques & Items by category in actor sheet (pc & npc) for better readability
+- Armor & Weapon added in the conflict tab now set the "eqquiped" props by default
 
 ## 1.1.2 - One Compendium to bring them all
 - Added compendiums (Thanks to Stéfano Fara for the English version !) Partial for French as PoW and CR are not translated yet